+++ libopendmi/src/entity/group-assoc.c
@@ -11,6 +11,8 @@
 #include <opendmi/entity/group-assoc.h>
 
 static bool dmi_group_assoc_decode(dmi_entity_t *entity);
+static bool dmi_group_assoc_link(dmi_entity_t *entity);
+static void dmi_group_assoc_cleanup(dmi_entity_t *entity);
 
 const dmi_entity_spec_t dmi_group_assoc_spec =
 {
@@ -18,45 +20,97 @@ const dmi_entity_spec_t dmi_group_assoc_spec =
     .name            = "Group associations",
     .type            = DMI_TYPE_GROUP_ASSOC,
     .minimum_version = DMI_VERSION(2, 0, 0),
-    .minimum_length  = 0x08,
+    .minimum_length  = 0x05,
     .decoded_length  = sizeof(dmi_group_assoc_t),
     .attributes      = (dmi_attribute_t[]){
         DMI_ATTRIBUTE(dmi_group_assoc_t, group_name, STRING, {
-            .code = "group-name",
-            .name = "Group name"
+            .code  = "group-name",
+            .name  = "Group name"
         }),
-        DMI_ATTRIBUTE(dmi_group_assoc_t, item_type, INTEGER, {
-            .code  = "item-type",
-            .name  = "Item type",
-            .flags = DMI_ATTRIBUTE_FLAG_HEX
-        }),
-        DMI_ATTRIBUTE(dmi_group_assoc_t, item_handle, HANDLE, {
-            .code = "item-handle",
-            .name = "Item handle"
+        DMI_ATTRIBUTE_ARRAY(dmi_group_assoc_t, items, item_count, STRUCT, {
+            .code  = "items",
+            .name  = "Items",
+            .attrs = (dmi_attribute_t[]){
+                DMI_ATTRIBUTE(dmi_group_assoc_item_t, type, INTEGER, {
+                    .code  = "type",
+                    .name  = "Type",
+                    .flags = DMI_ATTRIBUTE_FLAG_HEX
+                }),
+                DMI_ATTRIBUTE(dmi_group_assoc_item_t, handle, HANDLE, {
+                    .code = "Handle",
+                    .name = "Handle"
+                }),
+                DMI_ATTRIBUTE_NULL
+            }
         }),
         DMI_ATTRIBUTE_NULL
     },
     .handlers = {
-        .decode = dmi_group_assoc_decode,
+        .decode  = dmi_group_assoc_decode,
+        .link    = dmi_group_assoc_link,
+        .cleanup = dmi_group_assoc_cleanup
     }
 };
 
 static bool dmi_group_assoc_decode(dmi_entity_t *entity)
 {
     dmi_group_assoc_t *info;
-    const dmi_group_assoc_data_t *data;
-
-    data = dmi_entity_data(entity, DMI_TYPE_GROUP_ASSOC);
-    if (data == nullptr)
-        return false;
 
     info = dmi_entity_info(entity, DMI_TYPE_GROUP_ASSOC);
     if (info == nullptr)
         return false;
 
-    info->group_name  = dmi_entity_string(entity, data->group_name);
-    info->item_type   = dmi_decode(data->item_type);
-    info->item_handle = dmi_decode(data->item_handle);
+    dmi_stream_t *stream = &entity->stream;
+
+    if (not dmi_stream_decode_str(stream, &info->group_name))
+        return false;
+
+    info->item_count = (entity->body_length - 5) / 3;
+
+    info->items = dmi_alloc_array(entity->context, sizeof(dmi_group_assoc_item_t), info->item_count);
+    if (info->items == nullptr)
+        return false;
+
+    bool status = true;
+    for (size_t i = 0; i < info->item_count; i++) {
+        dmi_group_assoc_item_t *item = &info->items[i];
+
+        status =
+            dmi_stream_decode(stream, dmi_byte_t, &item->type) &&
+            dmi_stream_decode(stream, dmi_handle_t, &item->handle);
+
+        if (not status)
+            break;
+    }
+
+    return status;
+}
+
+static bool dmi_group_assoc_link(dmi_entity_t *entity)
+{
+    dmi_group_assoc_t *info;
+
+    info = dmi_entity_info(entity, DMI_TYPE_GROUP_ASSOC);
+    if (info == nullptr)
+        return false;
+
+    dmi_registry_t *registry = entity->context->registry;
+
+    for (size_t i = 0; i < info->item_count; i++) {
+        dmi_group_assoc_item_t *item = &info->items[i];
+        item->entity = dmi_registry_get(registry, item->handle, item->type, false);
+    }
 
     return true;
+}
+
+static void dmi_group_assoc_cleanup(dmi_entity_t *entity)
+{
+    dmi_group_assoc_t *info;
+
+    info = dmi_entity_info(entity, DMI_TYPE_GROUP_ASSOC);
+    if (info == nullptr)
+        return;
+
+    dmi_free(info->items);
 }