+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Fixed hardware security settings decoding
 - Fixed baseboard structure minumum length
+- Fixed group associations items list decoding
 
 ## [0.1.3] - January 11, 2026